+++ lib/app/features/feed/create_article/views/pages/create_article_modal/create_article_modal.dart
@@ -30,7 +30,7 @@ class CreateArticleModal extends StatelessWidget {
               ),
             ],
           ),
-          Expanded(child: SizedBox.shrink()),
+          Spacer(),
         ],
       ),
     );
