+++ lib/app/features/feed/create_story/views/pages/create_story_modal/create_story_modal.dart
@@ -3,8 +3,8 @@ import 'package:ice/app/extensions/extensions.dart';
 import 'package:ice/app/router/components/navigation_app_bar/navigation_app_bar.dart';
 import 'package:ice/app/router/components/sheet_content/sheet_content.dart';
 
-class CreateVideoModal extends StatelessWidget {
-  const CreateVideoModal({super.key});
+class CreateStoryModal extends StatelessWidget {
+  const CreateStoryModal({super.key});
 
   @override
   Widget build(BuildContext context) {
@@ -13,9 +13,9 @@ class CreateVideoModal extends StatelessWidget {
         mainAxisSize: MainAxisSize.min,
         children: [
           NavigationAppBar.modal(
-            title: Text(context.i18n.create_video_nav_title),
+            title: Text(context.i18n.create_story_nav_title),
           ),
-          Expanded(child: SizedBox.shrink()),
+          Spacer(),
         ],
       ),
     );
