+++ lib/app/features/feed/create_video/views/pages/create_video_modal/create_video_modal.dart
@@ -3,8 +3,8 @@ import 'package:ice/app/extensions/extensions.dart';
 import 'package:ice/app/router/components/navigation_app_bar/navigation_app_bar.dart';
 import 'package:ice/app/router/components/sheet_content/sheet_content.dart';
 
-class CreateStoryModal extends StatelessWidget {
-  const CreateStoryModal({super.key});
+class CreateVideoModal extends StatelessWidget {
+  const CreateVideoModal({super.key});
 
   @override
   Widget build(BuildContext context) {
@@ -13,9 +13,9 @@ class CreateStoryModal extends StatelessWidget {
         mainAxisSize: MainAxisSize.min,
         children: [
           NavigationAppBar.modal(
-            title: Text(context.i18n.create_story_nav_title),
+            title: Text(context.i18n.create_video_nav_title),
           ),
-          Expanded(child: SizedBox.shrink()),
+          Spacer(),
         ],
       ),
     );
